modified:   include/UltrasonicSensor.h 	modified:   src/CarController.cpp 	modified:   src/UltrasonicSensor.cpp 	modified:   src/main.cpp

+++ include/UltrasonicSensor.h
@@ -8,7 +8,7 @@ class UltrasonicSensor
 public:
     UltrasonicSensor(int trigPin, int echoPin);
     void init();
-    int getDistance();
+    float getDistance();
 
 private:
     int trigPin;
